+++ apps/decodex/src/orchestrator/tests/review_landing_status_rows/closeout_visibility.rs
@@ -1,0 +1,307 @@
+use crate::{
+	orchestrator::{
+		self, StateStore, tests,
+		tests::{FakePullRequestReviewStateInspector, FakeTracker, recovery_terminal_support},
+	},
+	test_support,
+	worktree::{WorktreeManager, WorktreeSpec},
+};
+
+#[test]
+fn build_post_review_lane_statuses_keeps_completed_issue_visible_for_closeout_tail_work() {
+	let (temp_dir, base_config, workflow) = tests::temp_project_layout();
+	let config = tests::service_config_with_github_token_env_var(&base_config, "HOME");
+	let repo_root = config.repo_root().to_path_buf();
+	let issue = tests::sample_issue("Done", &[]);
+	let tracker =
+		FakeTracker::with_refresh_snapshots(vec![issue.clone()], vec![vec![issue.clone()]]);
+	let state_store = StateStore::open_in_memory().expect("state store should open");
+	let head_oid = String::from_utf8(
+		test_support::hermetic_git_command()
+			.arg("-C")
+			.arg(&repo_root)
+			.args(["rev-parse", "HEAD"])
+			.output()
+			.expect("git rev-parse should run")
+			.stdout,
+	)
+	.expect("git output should be utf-8")
+	.trim()
+	.to_owned();
+	let pr_url = "https://github.com/hack-ink/decodex/pull/173";
+	let worktree_spec = WorktreeSpec {
+		branch_name: String::from("main"),
+		issue_identifier: issue.identifier.clone(),
+		path: repo_root.clone(),
+		reused_existing: true,
+	};
+	let _path_guard = recovery_terminal_support::install_fake_closeout_gh_responses(
+		&temp_dir,
+		&worktree_spec,
+		pr_url,
+		&head_oid,
+	);
+
+	state_store
+		.upsert_worktree("pubfi", &issue.id, "main", &repo_root.display().to_string())
+		.expect("worktree should record");
+
+	tests::seed_review_handoff_marker_for_path(
+		&state_store,
+		config.service_id(),
+		&repo_root,
+		&tests::sample_review_handoff_marker("main", pr_url, &head_oid),
+	);
+
+	let mut review_state = tests::sample_pull_request_review_state(
+		pr_url,
+		"main",
+		&head_oid,
+		Some("APPROVED"),
+		"MERGEABLE",
+		"CLEAN",
+		Some("SUCCESS"),
+		0,
+	);
+
+	review_state.state = String::from("MERGED");
+
+	let lanes = orchestrator::build_post_review_lane_statuses(
+		&tracker,
+		&config,
+		&workflow,
+		&state_store,
+		&FakePullRequestReviewStateInspector::new(vec![Ok(review_state)]),
+	)
+	.expect("post-review lane status build should succeed");
+
+	assert_eq!(lanes.len(), 1);
+	assert_eq!(lanes[0].issue_state, "Done");
+	assert_eq!(lanes[0].classification, "continue");
+	assert_eq!(lanes[0].reason, "pull_request_merged_closeout_pending");
+	assert_eq!(lanes[0].pr_url.as_deref(), Some(pr_url));
+}
+
+#[test]
+fn build_post_review_lane_statuses_keeps_merged_closeout_visible_after_retry_budget() {
+	let (temp_dir, base_config, workflow) = tests::temp_project_layout();
+	let config = tests::service_config_with_github_token_env_var(&base_config, "HOME");
+	let repo_root = config.repo_root().to_path_buf();
+	let issue = tests::sample_issue("In Review", &[]);
+	let tracker =
+		FakeTracker::with_refresh_snapshots(vec![issue.clone()], vec![vec![issue.clone()]]);
+	let state_store = StateStore::open_in_memory().expect("state store should open");
+	let head_oid = String::from_utf8(
+		test_support::hermetic_git_command()
+			.arg("-C")
+			.arg(&repo_root)
+			.args(["rev-parse", "HEAD"])
+			.output()
+			.expect("git rev-parse should run")
+			.stdout,
+	)
+	.expect("git output should be utf-8")
+	.trim()
+	.to_owned();
+	let pr_url = "https://github.com/hack-ink/decodex/pull/173";
+	let worktree_spec = WorktreeSpec {
+		branch_name: String::from("main"),
+		issue_identifier: issue.identifier.clone(),
+		path: repo_root.clone(),
+		reused_existing: true,
+	};
+	let _path_guard = recovery_terminal_support::install_fake_closeout_gh_responses(
+		&temp_dir,
+		&worktree_spec,
+		pr_url,
+		&head_oid,
+	);
+
+	state_store
+		.upsert_worktree("pubfi", &issue.id, "main", &repo_root.display().to_string())
+		.expect("worktree should record");
+
+	tests::seed_review_handoff_marker_for_path(
+		&state_store,
+		config.service_id(),
+		&repo_root,
+		&tests::sample_review_handoff_marker("main", pr_url, &head_oid),
+	);
+
+	for attempt in 1..=3 {
+		state_store
+			.record_run_attempt(&format!("run-{attempt}"), &issue.id, attempt, "failed")
+			.expect("failed attempt should record");
+	}
+
+	let mut review_state = tests::sample_pull_request_review_state(
+		pr_url,
+		"main",
+		&head_oid,
+		Some("APPROVED"),
+		"MERGEABLE",
+		"CLEAN",
+		Some("SUCCESS"),
+		0,
+	);
+
+	review_state.state = String::from("MERGED");
+
+	let lanes = orchestrator::build_post_review_lane_statuses(
+		&tracker,
+		&config,
+		&workflow,
+		&state_store,
+		&FakePullRequestReviewStateInspector::new(vec![Ok(review_state)]),
+	)
+	.expect("post-review lane status build should succeed");
+
+	assert_eq!(lanes.len(), 1);
+	assert_eq!(lanes[0].issue_state, "In Review");
+	assert_eq!(lanes[0].classification, "continue");
+	assert_eq!(lanes[0].reason, "pull_request_merged_closeout_pending");
+	assert_eq!(lanes[0].pr_url.as_deref(), Some(pr_url));
+	assert_eq!(lanes[0].pr_state.as_deref(), Some("MERGED"));
+}
+
+#[test]
+fn build_post_review_lane_statuses_keeps_merged_closeout_visible_after_landed_main_advances() {
+	let (temp_dir, base_config, workflow) = tests::temp_project_layout();
+	let config = tests::service_config_with_github_token_env_var(&base_config, "HOME");
+	let issue = tests::sample_issue("In Review", &[]);
+	let tracker =
+		FakeTracker::with_refresh_snapshots(vec![issue.clone()], vec![vec![issue.clone()]]);
+	let state_store = StateStore::open_in_memory().expect("state store should open");
+	let worktree_manager =
+		WorktreeManager::new(config.service_id(), config.repo_root(), config.worktree_root());
+	let worktree =
+		worktree_manager.ensure_worktree(&issue.identifier, false).expect("worktree should exist");
+	let pr_head_oid = tests::git_output(&worktree.path, &["rev-parse", "HEAD"]);
+	let merge_commit_oid = tests::commit_worktree_change(
+		&worktree.path,
+		"landed.txt",
+		"landed\n",
+		"land retained lane",
+	);
+	let current_head_oid =
+		tests::commit_worktree_change(&worktree.path, "later.txt", "later\n", "advance main later");
+	let pr_url = "https://github.com/hack-ink/decodex/pull/203";
+	let _path_guard = recovery_terminal_support::install_fake_closeout_gh_responses(
+		&temp_dir,
+		&worktree,
+		pr_url,
+		&pr_head_oid,
+	);
+
+	state_store
+		.upsert_worktree(
+			config.service_id(),
+			&issue.id,
+			&worktree.branch_name,
+			&worktree.path.display().to_string(),
+		)
+		.expect("worktree should record");
+
+	tests::seed_review_handoff_marker_for_path(
+		&state_store,
+		config.service_id(),
+		&worktree.path,
+		&tests::sample_review_handoff_marker(&worktree.branch_name, pr_url, &pr_head_oid),
+	);
+
+	let mut review_state = tests::sample_pull_request_review_state(
+		pr_url,
+		&worktree.branch_name,
+		&pr_head_oid,
+		Some("APPROVED"),
+		"MERGEABLE",
+		"CLEAN",
+		Some("SUCCESS"),
+		0,
+	);
+
+	review_state.state = String::from("MERGED");
+	review_state.merge_commit_oid = Some(merge_commit_oid);
+
+	let lanes = orchestrator::build_post_review_lane_statuses(
+		&tracker,
+		&config,
+		&workflow,
+		&state_store,
+		&FakePullRequestReviewStateInspector::new(vec![Ok(review_state)]),
+	)
+	.expect("post-review lane status build should succeed");
+
+	assert_eq!(tests::git_output(&worktree.path, &["rev-parse", "HEAD"]), current_head_oid);
+	assert_eq!(lanes.len(), 1);
+	assert_eq!(lanes[0].classification, "continue");
+	assert_eq!(lanes[0].reason, "pull_request_merged_closeout_pending");
+	assert_eq!(lanes[0].pr_url.as_deref(), Some(pr_url));
+	assert_eq!(lanes[0].pr_state.as_deref(), Some("MERGED"));
+}
+
+#[test]
+fn build_post_review_lane_statuses_blocks_closeout_when_merge_readbacks_conflict() {
+	let (temp_dir, base_config, workflow) = tests::temp_project_layout();
+	let config = tests::service_config_with_github_token_env_var(&base_config, "HOME");
+	let issue = tests::sample_issue("In Review", &[]);
+	let tracker =
+		FakeTracker::with_refresh_snapshots(vec![issue.clone()], vec![vec![issue.clone()]]);
+	let state_store = StateStore::open_in_memory().expect("state store should open");
+	let worktree_manager =
+		WorktreeManager::new(config.service_id(), config.repo_root(), config.worktree_root());
+	let worktree = worktree_manager
+		.ensure_worktree(&issue.identifier, false)
+		.expect("retained worktree should exist");
+	let head_oid = tests::git_output(&worktree.path, &["rev-parse", "HEAD"]);
+	let pr_url = "https://github.com/hack-ink/decodex/pull/204";
+	let _path_guard = recovery_terminal_support::install_fake_closeout_gh_responses_with_states(
+		&temp_dir, &worktree, pr_url, &head_oid, "OPEN", "MERGED",
+	);
+
+	state_store
+		.upsert_worktree(
+			config.service_id(),
+			&issue.id,
+			&worktree.branch_name,
+			&worktree.path.display().to_string(),
+		)
+		.expect("worktree should record");
+
+	tests::seed_review_handoff_marker_for_path(
+		&state_store,
+		config.service_id(),
+		&worktree.path,
+		&tests::sample_review_handoff_marker(&worktree.branch_name, pr_url, &head_oid),
+	);
+
+	let mut review_state = tests::sample_pull_request_review_state(
+		pr_url,
+		&worktree.branch_name,
+		&head_oid,
+		Some("APPROVED"),
+		"MERGEABLE",
+		"CLEAN",
+		Some("SUCCESS"),
+		0,
+	);
+
+	review_state.state = String::from("MERGED");
+
+	let lanes = orchestrator::build_post_review_lane_statuses(
+		&tracker,
+		&config,
+		&workflow,
+		&state_store,
+		&FakePullRequestReviewStateInspector::new(vec![Ok(review_state)]),
+	)
+	.expect("post-review lane status build should succeed");
+
+	assert_eq!(lanes.len(), 1);
+	assert_eq!(lanes[0].classification, "blocked");
+	assert_eq!(lanes[0].reason, "pull_request_merge_state_conflict");
+	assert_eq!(lanes[0].readback_warning.as_deref(), Some("pull_request_merge_state_conflict"));
+	assert_eq!(lanes[0].readback_root_cause.as_deref(), Some("lineage_validation_failed"));
+	assert_eq!(lanes[0].pr_state.as_deref(), Some("OPEN"));
+	assert_eq!(lanes[0].pr_url.as_deref(), Some(pr_url));
+}
