+++ apps/decodex/src/orchestrator/tests/review_landing_status_rows/external_review.rs
@@ -1,0 +1,141 @@
+use crate::{
+	orchestrator::{
+		self, StateStore, tests,
+		tests::{
+			FakePullRequestReviewStateInspector, FakeTracker, review_landing_status_rows::support,
+		},
+	},
+	test_support,
+};
+
+#[test]
+fn build_post_review_lane_statuses_reports_ready_to_land() {
+	let (_temp_dir, config, workflow) = tests::temp_project_layout();
+	let repo_root = config.repo_root().to_path_buf();
+	let issue = tests::sample_issue("In Review", &[]);
+	let tracker =
+		FakeTracker::with_refresh_snapshots(vec![issue.clone()], vec![vec![issue.clone()]]);
+	let state_store = StateStore::open_in_memory().expect("state store should open");
+	let head_oid = String::from_utf8(
+		test_support::hermetic_git_command()
+			.arg("-C")
+			.arg(&repo_root)
+			.args(["rev-parse", "HEAD"])
+			.output()
+			.expect("git rev-parse should run")
+			.stdout,
+	)
+	.expect("git output should be utf-8")
+	.trim()
+	.to_owned();
+	let pr_url = "https://github.com/hack-ink/decodex/pull/173";
+
+	state_store
+		.upsert_worktree("pubfi", &issue.id, "main", &repo_root.display().to_string())
+		.expect("worktree should record");
+
+	tests::seed_review_handoff_marker_for_path(
+		&state_store,
+		config.service_id(),
+		&repo_root,
+		&tests::sample_review_handoff_marker("main", pr_url, &head_oid),
+	);
+	tests::seed_review_orchestration_marker_for_path(
+		&state_store,
+		config.service_id(),
+		&repo_root,
+		&tests::sample_review_orchestration_marker(
+			"main",
+			pr_url,
+			&head_oid,
+			"waiting_for_result",
+			1,
+		),
+	);
+
+	let mut review_state = tests::sample_pull_request_review_state(
+		pr_url,
+		"main",
+		&head_oid,
+		Some("APPROVED"),
+		"MERGEABLE",
+		"CLEAN",
+		Some("SUCCESS"),
+		0,
+	);
+
+	tests::add_external_review_ack(&mut review_state);
+	tests::add_external_review_pass(&mut review_state);
+
+	let lanes = orchestrator::build_post_review_lane_statuses(
+		&tracker,
+		&config,
+		&workflow,
+		&state_store,
+		&FakePullRequestReviewStateInspector::new(vec![Ok(review_state)]),
+	)
+	.expect("post-review lane status build should succeed");
+
+	assert_eq!(lanes.len(), 1);
+	assert_eq!(lanes[0].classification, "ready_to_land");
+	assert_eq!(lanes[0].reason, "external_review_passed_strict");
+	assert_eq!(lanes[0].pr_url.as_deref(), Some(pr_url));
+	assert!(
+		lanes[0].loop_status.as_ref().and_then(|status| status.review.as_ref()).is_none(),
+		"ready_to_land must not project a pending review checkpoint"
+	);
+	assert_eq!(lanes[0].readback_warning.as_deref(), Some("active_ownership_label_missing"));
+}
+
+#[test]
+fn build_post_review_lane_statuses_waits_when_clean_repair_head_outruns_lifecycle_marker() {
+	let (_temp_dir, config, workflow) = tests::temp_project_layout();
+	let issue = tests::sample_issue("In Review", &[]);
+	let tracker =
+		FakeTracker::with_refresh_snapshots(vec![issue.clone()], vec![vec![issue.clone()]]);
+	let state_store = StateStore::open_in_memory().expect("state store should open");
+	let pr_url = "https://github.com/hack-ink/decodex/pull/173";
+	let (worktree, repaired_head_oid) =
+		support::retained_worktree_with_stale_review_lifecycle_marker(
+			&config,
+			&state_store,
+			&issue,
+			pr_url,
+		);
+
+	support::seed_clean_repair_completion_writeback_gap(
+		&config,
+		&state_store,
+		&issue,
+		&worktree,
+		pr_url,
+		&repaired_head_oid,
+	);
+
+	let review_state = tests::sample_pull_request_review_state(
+		pr_url,
+		&worktree.branch_name,
+		&repaired_head_oid,
+		Some("APPROVED"),
+		"MERGEABLE",
+		"CLEAN",
+		Some("SUCCESS"),
+		0,
+	);
+	let lanes = orchestrator::build_post_review_lane_statuses(
+		&tracker,
+		&config,
+		&workflow,
+		&state_store,
+		&FakePullRequestReviewStateInspector::new(vec![Ok(review_state)]),
+	)
+	.expect("post-review lane status build should succeed");
+
+	assert_eq!(lanes.len(), 1);
+	assert_eq!(lanes[0].classification, "wait_for_review");
+	assert_eq!(lanes[0].reason, "review_repair_writeback_stale_lifecycle_marker");
+	assert_eq!(
+		lanes[0].readback_warning.as_deref(),
+		Some("review_repair_writeback_stale_lifecycle_marker")
+	);
+}
